+++ api-server/package.json
@@ -4,13 +4,14 @@
   "main": "index.js",
   "scripts": {
     "watch": "tsc -w",
-    "dev": "nodemon dist/core/index.js",
-    "start": "node dist/core/index.js",
-    "fake-dev": "nodemon --exec ts-node src/core/index.ts",
-    "fake-start": "ts-node src/core/index.ts"
+    "dev": "nodemon dist/core/main.js",
+    "start": "node dist/core/main.js",
+    "fake-dev": "nodemon --exec ts-node src/core/main.ts",
+    "fake-start": "ts-node src/core/main.ts"
   },
   "license": "MIT",
   "devDependencies": {
+    "@types/express": "^4.17.11",
     "@types/node": "^15.3.0",
     "nodemon": "^2.0.7",
     "ts-node": "^9.1.1",
@@ -21,7 +22,12 @@
     "@mikro-orm/core": "^4.5.5",
     "@mikro-orm/migrations": "^4.5.5",
     "@mikro-orm/mongodb": "^4.5.5",
-    "mongodb": "^3.6.6"
+    "apollo-server-express": "^2.24.0",
+    "class-validator": "^0.13.1",
+    "express": "^4.17.1",
+    "graphql": "^15.5.0",
+    "mongodb": "^3.6.6",
+    "type-graphql": "^1.1.1"
   },
   "mikro-orm": {
     "useTsNode": true,
