+++ api-server/src/entities/index.ts
@@ -1,0 +1,2 @@
+export * from "./post";
+export * from "./mikro-orm.config";
