+++ api-server/src/resolvers/hello.ts
@@ -1,0 +1,9 @@
+import { Query, Resolver } from "type-graphql";
+
+@Resolver()
+export class HelloResolver {
+	@Query(() => String)
+	hello() {
+		return "bonjour";
+	}
+}
